+++ DependencyInjection/CmfSearchExtension.php
@@ -2,10 +2,11 @@
 
 namespace Symfony\Cmf\Bundle\SearchBundle\DependencyInjection;
 
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\Config\FileLocator;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
+use Symfony\Component\Config\FileLocator;
 
 class CmfSearchExtension extends Extension
 {
@@ -17,21 +18,19 @@ class CmfSearchExtension extends Extension
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
 
         $config = $this->processConfiguration(new Configuration(), $configs);
-        $container->setParameter($this->getAlias().'.document_manager_name', $config['document_manager_name']);
-        $searchPath = $config['search_path'];
-        if (null === $searchPath) {
-            if ($container->hasParameter('cmf_core.content_basepath')) {
-                $searchPath = $container->getParameter('cmf_core.content_basepath');
-            } else {
-                $searchPath = '/cms/content';
-            }
-        }
-        $container->setParameter($this->getAlias() . '.search_path', $searchPath);
-        $container->setParameter($this->getAlias() . '.search_fields', $config['search_fields']);
-        $container->setParameter($this->getAlias() . '.translation_strategy', $config['translation_strategy']);
         $container->setParameter($this->getAlias() . '.show_paging', $config['show_paging']);
 
-        $loader->load('services.xml');
+        if ($config['persistence']['phpcr']) {
+            $loader->load('services.xml');
+
+            $searchController = $container->getDefinition($this->getAlias().'.phpcr_controller');
+            $searchController->replaceArgument(0, new Reference($config['persistence']['phpcr']['manager_registry']));
+
+            $container->setParameter($this->getAlias() . '.persistence.phpcr.manager_name', $config['persistence']['phpcr']['manager_name']);
+            $container->setParameter($this->getAlias() . '.persistence.phpcr.search_basepath', $config['persistence']['phpcr']['search_basepath']);
+            $container->setParameter($this->getAlias() . '.persistence.phpcr.search_fields', $config['persistence']['phpcr']['search_fields']);
+            $container->setParameter($this->getAlias() . '.persistence.phpcr.translation_strategy', $config['persistence']['phpcr']['translation_strategy']);
+        }
     }
 
     /**