+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+* **2013-09-04**: [DependencyInjection] restructured config into `persistence` -> `phpcr` to match other Bundles
+
 1.0.0-beta2
 -----------